pyproject.toml
[project]
name = "nfl-predictions-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
# readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.38.0,<0.39.0)",
    "pymongo (>=4.15.3,<5.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.121.1,<0.122.0)",
    "pytest (>=8.2.0,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
